feat: BasicButton 및 GhostButton 컴포넌트 타입 통합 및 불필요한 파일 삭제
