--- frontend/src/components/BasicButton/BasicButton.types.ts
@@ -1,6 +1,0 @@
-export interface BasicButtonProps extends React.ComponentProps<'button'> {
-  children: React.ReactNode;
-  width?: string | number;
-  icon?: React.ReactNode;
-  variant?: 'primary' | 'secondary';
-}
