--- frontend/src/components/GhostButton/GhostButton.types.ts
@@ -1,4 +1,0 @@
-export interface GhostButtonProps extends React.ComponentProps<'button'> {
-  icon: React.ReactNode;
-  text: string;
-}
