--- frontend/src/domains/user/home/components/Button/Button.tsx
@@ -1,16 +1,0 @@
-import { css } from '@emotion/react';
-
-export default function LikeButton() {
-  return (
-    <div
-      css={css`
-        position: absolute;
-        width: 100px;
-        height: 100px;
-        background-color: aliceblue;
-      `}
-    >
-      hello
-    </div>
-  );
-}
