--- frontend/src/domains/user/home/components/Hero/Hero.types.ts
@@ -1,5 +1,0 @@
-export interface HeroProps extends React.ComponentProps<'div'> {
-  onLoginClick: () => void;
-  onSuggestClick: () => void;
-  title: string;
-}
